backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.4
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
# black==24.10.0
# ruff==0.9.6
